+++ pkg/yarus_platform/reference/domain/entity_type2property/repository.go
@@ -14,7 +14,7 @@ type Repository interface {
 	First(ctx context.Context, entity *EntityType2Property) (*EntityType2Property, error)
 	// Query returns the list of albums with the given offset and limit.
 	Query(ctx context.Context, cond *selection_condition.SelectionCondition) ([]EntityType2Property, error)
-	Count(ctx context.Context, cond *selection_condition.SelectionCondition) (uint, error)
+	Count(ctx context.Context, cond *selection_condition.SelectionCondition) (int64, error)
 	Create(ctx context.Context, entity *EntityType2Property) error
 	Update(ctx context.Context, entity *EntityType2Property) error
 	Save(ctx context.Context, entity *EntityType2Property) error
