+++ pkg/yarus_platform/reference/domain/text_value/repository.go
@@ -14,7 +14,7 @@ type Repository interface {
 	First(ctx context.Context, entity *TextValue) (*TextValue, error)
 	// Query returns the list of albums with the given offset and limit.
 	Query(ctx context.Context, cond *selection_condition.SelectionCondition) ([]TextValue, error)
-	Count(ctx context.Context, cond *selection_condition.SelectionCondition) (uint, error)
+	Count(ctx context.Context, cond *selection_condition.SelectionCondition) (int64, error)
 	Create(ctx context.Context, entity *TextValue) error
 	Update(ctx context.Context, entity *TextValue) error
 	Save(ctx context.Context, entity *TextValue) error
